Add ability to mark tasks as done or not
There is no way to keep the list updated with the current status of a
task. Task as a list of strings is too limited as a data type.

Let's keep tasks as a class on it's own and add isDone boolean value
to it, and add new commands to mark and unmark tasks. Add in a method
to handle out of bound demands to chatbot.