Add error handling for event and deadline commands
Invalid or empty strings can be entered by the user for deadline and
event command, which causes the code to crash. The main() function
has too many try-catch blocks, impairs readability of code.

Let's add in error-handling for empty or invalid command syntax for
user, and let's abstract out error-handling to the handle functions.